requirements.txt
ollama
pydantic>=2.0
httpx>=0.28
requests>=2.32
rich
python-dotenv
